+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# emphatic 0.1.6.9003  2024-04-28
+# emphatic 0.1.6.9004  2024-04-30
 
 * Removed `hl_mat()` and associated vignettes.
     * function is not exported for now as the selection process is too 
@@ -9,6 +9,7 @@
 * Refactored arguments to `hl()` for more clarity
 * Added latex output so Quarto/Rmd will show emphatic objects when 
   rendered to PDF
+* Remove `scale_mode` argument to `hl()`
 
 # emphatic 0.1.6  2024-04-27